requirements.txt
textstat
lexical-diversity
vaderSentiment
vaderSentiment
transformers
sentencepiece
tensorflow-addons
tensorflow_text